% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_bar}
\alias{color_bar}
\title{color_bar
Plot a colorbar "legend".}
\usage{
color_bar(lut, ticks, Title = "", cex_legend_labels = 1)
}
\arguments{
\item{lut}{A vector of colors}

\item{ticks}{The number of ticks in the legend}

\item{Title}{The title of the scale bar}

\item{cex_legend_labels}{The size of the lables for the color_bar}
}
\value{
Color bar plot with ticks
}
\description{
color_bar
Plot a colorbar "legend".
}
