% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibes_load.R
\name{sibes_load}
\alias{sibes_load}
\title{sibes_load}
\usage{
sibes_load(
  directory_path,
  biomass_csv_name = "biota.csv",
  species_csv_name = "species.csv",
  samples_csv_name = "samples.csv"
)
}
\arguments{
\item{directory_path}{The path to the folder containing the csv files}

\item{biomass_csv_name}{String with the name of the csv file 'biota'}

\item{species_csv_name}{String with the name of the csv file species}

\item{samples_csv_name}{character vector name of csv file 'samples'}
}
\value{
A list containing three tibbles with the data
}
\description{
This function allow an easy load of the three csv-files with SIBES data
}
