% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibes_plot_map.R
\name{sibes_plot_map}
\alias{sibes_plot_map}
\title{sibes_plot_map}
\usage{
sibes_plot_map(
  Inputdata,
  attribute,
  ticks = 5,
  InputdataProj = 4326,
  MapProj = 28991,
  cex_map = 0.2,
  cex_legend_labels = 1
)
}
\arguments{
\item{Inputdata}{A tibble or data.frame containing "x" and "y" coordinates and the data to plot}

\item{attribute}{The name of the column with data to plot}

\item{InputdataProj}{CRS format of "x" and "y" in the data
EPSG:4326     (The standard latitude/longitude projection based on the WGS84 datum)
EPSG:28992     (The standard Amersfoort / RD New -- Netherlands - Holland - Dutch projection based on the WGS84 datum)
EPSG:32631     (WGS 84 / UTM zone 31N Between 0°E and 6°E, northern hemisphere between equator and 84°N, onshore and offshore. Algeria. Andorra. Belgium. Benin. Burkina Faso. Denmark - North Sea. France. Germany - North Sea. Ghana. Luxembourg. Mali. Netherlands. Niger. Nigeria. Norway. Spain. Togo. United Kingdom (UK) - North Sea.)}

\item{MapProj}{CRS format of the map plot}
}
\value{
A plotted map and color scale bar
}
\description{
sibes_plot_map
}
