% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibes_species.R
\name{sibes_species_search}
\alias{sibes_species_search}
\title{sibes_species_search
This function use keywords to search the species tibble. This is convenient if only parts of a name are known, or for example, when multiple species need to be selected with a shared keyword}
\usage{
sibes_species_search(Inputdata = NULL, Key = NULL)
}
\arguments{
\item{Inputdata}{Input data for the function is the output format of sibes_load which is data list of tibbels including "species". If these names are not present, the function returns an error}

\item{Key}{The keyword to search for as a string. Numeric values also entered as character}
}
\value{
Function returns the found search.
}
\description{
sibes_species_search
This function use keywords to search the species tibble. This is convenient if only parts of a name are known, or for example, when multiple species need to be selected with a shared keyword
}
\examples{

data_list <- get('SIBES_dataset')

sibes_species_search(data_list[["species"]], Key='57|34|67')
}
